Added beforeClose and beforeOpen callbacks
If you want e.g. change the background of the menu when it´s open, you would normally do it with the open callback. But then the background changes AFTER the menu is opened. So the user got a flickering of the change for the duration of the animation. So I added the beforeClose and the beforeOpen callback.